requirements.txt
# Core streaming and data processing
apache-kafka-python==2.0.2
kafka-python==2.0.2
confluent-kafka==2.3.0

# Database drivers and connectors
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Redis for caching
redis==5.0.1
redis-py-cluster==2.1.3

# Google Cloud BigQuery
google-cloud-bigquery==3.14.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Data manipulation and validation
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Async processing
asyncio-mqtt==0.16.1
asyncpg==0.29.0
aioredis==2.0.1

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# HTTP clients
httpx==0.25.2
requests==2.31.0

# JSON and serialization
orjson==3.9.10
jsonschema==4.20.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Time handling
pytz==2023.3
python-dateutil==2.8.2

# Health checks and utilities
tenacity==8.2.3
schedule==1.2.0
